# StayLocal_OpenAPI.yaml
openapi: 3.0.3
info:
  title: StayLocal Backend API
  version: 1.0.0
  description: Contract-first REST API for the StayLocal community tourism platform.
    Covers authentication, destinations, properties, media, camera+GPS verification,
    bookings, rewards, reviews and admin moderation.
  contact:
    name: StayLocal Project Team
    email: staylocal-team@example.com
servers:
- url: http://localhost:8787
  description: Local development
- url: http://localhost:8787
  description: Demo/CI default
...
paths:
...
  /api/v1/properties:
    get:
      operationId: listProperties
      summary: Search approved properties
      description: Returns public, approved listings with price and destination filters
        and optional date availability filtering. Results are paginated.
      tags:
      - Properties
      security: *id001
      parameters:
      - name: destinationId
        in: query
        required: false
        schema:
          type: string
        examples:
          200_properties_filtered:
            value: dest-bhimashankar
      - name: minPrice
        in: query
        required: false
        schema:
          type: number
          minimum: 0
        examples:
          200_properties_filtered:
            value: 1000
      - name: maxPrice
        in: query
        required: false
        schema:
          type: number
          minimum: 0
        examples:
          200_properties_filtered:
            value: 3000
      - name: checkIn
        in: query
        required: false
        schema:
          type: string
          format: date
        examples:
          200_properties_available:
            value: '2026-10-10'
      - name: checkOut
        in: query
        required: false
        schema:
          type: string
          format: date
        examples:
          200_properties_available:
            value: '2026-10-12'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
        examples:
          200_properties_paged:
            value: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 12
        examples:
          200_properties_paged:
            value: 12
      responses:
        '200':
          description: Property results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProperties'
              examples:
                200_properties_filtered:
                  value:
                    items:
                    - &id005
                      id: prop-1001
                      hostId: host-1001
                      destinationId: dest-bhimashankar
                      title: Forest View Local Homestay
                      description: Simple private room with a local host near Bhimashankar.
                      pricePerNight: 1800
                      maxGuests: 3
                      amenities:
                      - Wi-Fi
                      - Parking
                      - Hot Water
                      photos: []
                      status: APPROVED
                      latitude: 19.0731
                      longitude: 73.5356
                      locationVerified: true
                      verification: null
                      createdAt: '2026-09-20T09:00:00Z'
                      updatedAt: '2026-09-20T09:00:00Z'
                    page: 1
                    pageSize: 12
                    total: 1
        '422':
          description: One or more search parameters are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_properties_invalid_filter:
                  value:
                    code: VALIDATION_ERROR
                    message: One or more search parameters are invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '429':
          description: Request rate exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                429_properties_rate_limited:
                  value:
                    code: RATE_LIMITED
                    message: Request rate exceeded.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                500_properties_server_error:
                  value:
                    code: INTERNAL_ERROR
                    message: Unexpected server error.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
    post:
      operationId: createProperty
      summary: Create a room listing
      description: Creates a draft listing for an approved host. A non-approved host
        cannot publish or submit a listing.
      tags:
      - Properties
      security: *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - destinationId
              - title
              - description
              - pricePerNight
              - maxGuests
              - latitude
              - longitude
              properties:
                destinationId:
                  type: string
                title:
                  type: string
                  minLength: 3
                  maxLength: 160
                description:
                  type: string
                  minLength: 20
                  maxLength: 5000
                pricePerNight:
                  type: number
                  minimum: 1
                  maximum: 100000
                maxGuests:
                  type: integer
                  minimum: 1
                  maximum: 20
                amenities:
                  type: array
                  items:
                    type: string
                  maxItems: 30
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
            examples:
              201_property_create_success:
                value:
                  destinationId: dest-bhimashankar
                  title: Forest View Local Homestay
                  description: Simple private room with a local host near Bhimashankar.
                  pricePerNight: 1800
                  maxGuests: 3
                  amenities:
                  - Wi-Fi
                  - Parking
                  - Hot Water
                  latitude: 19.0731
                  longitude: 73.5356
              422_property_create_invalid_price:
                value:
                  destinationId: dest-bhimashankar
                  title: Bad Property
                  description: This description is long enough for validation.
                  pricePerNight: 0
                  maxGuests: 3
                  latitude: 19.0731
                  longitude: 73.5356
      responses:
        '201':
          description: Property created as DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
              examples:
                201_property_create_success:
                  value: *id005
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_property_create_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: Only an approved host can create listings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_property_create_forbidden:
                  value:
                    code: FORBIDDEN
                    message: Only an approved host can create listings.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_property_create_invalid_price:
                  value:
                    code: VALIDATION_ERROR
                    message: Request validation failed.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                500_property_create_server_error:
                  value:
                    code: INTERNAL_ERROR
                    message: Unexpected server error.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
...
  /api/v1/admin/hosts:
    get:
      operationId: listAdminHosts
      summary: List hosts for admin review
      description: Returns all host profiles with their approval state.
      tags:
      - Admin
      security: *id003
      responses:
        '200':
          description: Host list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HostProfile'
              examples:
                200_admin_hosts_success:
                  value:
                  - &id007
                    id: host-1001
                    userId: 20000000-0000-4000-8000-000000000001
                    displayName: Kiran Patil
                    bio: Local host near Bhimashankar.
                    approvalStatus: APPROVED
                    verifiedAt: '2026-09-20T09:00:00Z'
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_admin_hosts_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: Admin role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_admin_hosts_forbidden:
                  value:
                    code: FORBIDDEN
                    message: Admin role required.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/admin/hosts/{hostId}/status:
    patch:
      operationId: updateHostStatus
      summary: Approve or reject a host
      description: Admin-only action that changes host approval state. Rejection may
        be accompanied by an admin note.
      tags:
      - Admin
      security: *id003
      parameters:
      - name: hostId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - APPROVED
                  - REJECTED
                note:
                  type: string
                  maxLength: 1000
                  nullable: true
            examples:
              200_admin_host_approved:
                value:
                  status: APPROVED
                  note: Identity and profile reviewed.
              422_admin_host_invalid_status:
                value:
                  status: PENDING
      responses:
        '200':
          description: Host status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostProfile'
              examples:
                200_admin_host_approved:
                  value: *id007
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_admin_host_status_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: Admin role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_admin_host_status_forbidden:
                  value:
                    code: FORBIDDEN
                    message: Admin role required.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '404':
          description: Host not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_admin_host_status_missing:
                  value:
                    code: NOT_FOUND
                    message: Host not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Invalid host status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_admin_host_invalid_status:
                  value:
                    code: VALIDATION_ERROR
                    message: Invalid host status.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/admin/properties:
    get:
      operationId: listAdminProperties
      summary: List properties for admin review
      description: Returns all properties, including drafts and rejected listings,
        for admin moderation.
      tags:
      - Admin
      security: *id003
      responses:
        '200':
          description: Admin property list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
              examples:
                200_admin_properties_success:
                  value:
                  - *id005
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_admin_properties_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: Admin role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_admin_properties_forbidden:
                  value:
                    code: FORBIDDEN
                    message: Admin role required.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/admin/properties/{propertyId}/status:
    patch:
      operationId: updatePropertyStatus
      summary: Approve or reject a property
      description: Admin-only moderation action. Only approved properties are visible
        to public search and bookable.
      tags:
      - Admin
      security: *id003
      parameters:
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - APPROVED
                  - REJECTED
                  - INACTIVE
                note:
                  type: string
                  maxLength: 1000
                  nullable: true
            examples:
              200_admin_property_approved:
                value:
                  status: APPROVED
                  note: Listing reviewed.
              422_admin_property_invalid_status:
                value:
                  status: DRAFT
      responses:
        '200':
          description: Property status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
              examples:
                200_admin_property_approved:
                  value: *id005
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_admin_property_status_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: Admin role required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_admin_property_status_forbidden:
                  value:
                    code: FORBIDDEN
                    message: Admin role required.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_admin_property_status_missing:
                  value:
                    code: NOT_FOUND
                    message: Property not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Invalid property status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_admin_property_invalid_status:
                  value:
                    code: VALIDATION_ERROR
                    message: Invalid property status.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
components:
  schemas:
    StandardError:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          example: BOOKING_CONFLICT
        message:
          type: string
          example: The selected dates are already booked.
        requestId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        code: BOOKING_CONFLICT
        message: The selected dates are already booked.
        requestId: 11111111-1111-4111-8111-111111111111
        details:
          propertyId: prop-1001
...
    HostProfile:
      type: object
      required:
      - id
      - userId
      - displayName
      - approvalStatus
      properties:
        id:
          type: string
          example: host-1001
        userId:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        bio:
          type: string
          maxLength: 2000
          nullable: true
        approvalStatus:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        verifiedAt:
          type: string
          format: date-time
          nullable: true
      example: *id007
    Property:
      type: object
      required:
      - id
      - hostId
      - destinationId
      - title
      - description
      - pricePerNight
      - maxGuests
      - status
      - latitude
      - longitude
      - locationVerified
      - createdAt
      properties:
        id:
          type: string
          example: prop-1001
        hostId:
          type: string
          example: host-1001
        destinationId:
          type: string
          example: dest-bhimashankar
        title:
          type: string
          minLength: 3
          maxLength: 160
        description:
          type: string
          minLength: 20
          maxLength: 5000
        pricePerNight:
          type: number
          minimum: 1
          maximum: 100000
          example: 1800
        maxGuests:
          type: integer
          minimum: 1
          maximum: 20
          example: 3
        amenities:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 30
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PropertyPhoto'
          maxItems: 20
        status:
          type: string
          enum:
          - DRAFT
          - PENDING
          - APPROVED
          - REJECTED
          - INACTIVE
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        locationVerified:
          type: boolean
        verification:
          $ref: '#/components/schemas/VerificationResult'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example: *id005
    PropertyPhoto:
      type: object
      required:
      - id
      - assetUrl
      - isVerificationPhoto
      - createdAt
      properties:
        id:
          type: string
          example: photo-1001
        assetUrl:
          type: string
          format: uri
        caption:
          type: string
          maxLength: 200
          nullable: true
        isVerificationPhoto:
          type: boolean
        createdAt:
          type: string
          format: date-time
    VerificationResult:
      type: object
      required:
      - status
      - distanceMeters
      - thresholdMeters
      - capturedLatitude
      - capturedLongitude
      - propertyLatitude
      - propertyLongitude
      - capturedAt
      - mode
      - disclaimer
      properties:
        status:
          type: string
          enum:
          - VERIFIED
          - REJECTED
        distanceMeters:
          type: number
          minimum: 0
        thresholdMeters:
          type: number
          minimum: 1
          example: 50
        capturedLatitude:
          type: number
          minimum: -90
          maximum: 90
        capturedLongitude:
          type: number
          minimum: -180
          maximum: 180
        propertyLatitude:
          type: number
          minimum: -90
          maximum: 90
        propertyLongitude:
          type: number
          minimum: -180
          maximum: 180
        accuracyMeters:
          type: number
          minimum: 0
          nullable: true
        capturedAt:
          type: string
          format: date-time
        photoAssetId:
          type: string
        mode:
          type: string
          enum:
          - LIVE
          - SIMULATED_AT_PROPERTY
          - SIMULATED_AWAY
        disclaimer:
          type: string
          example: GPS proximity verifies device location at capture time; it does
            not independently guarantee room condition.
...
    PaginatedProperties:
      type: object
      required:
      - items
      - page
      - pageSize
      - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT